+++ src/cmosCitation.ts
@@ -11,20 +11,20 @@ import {
   eJournalBibliography, eJournalNoteList, journalBibliography, journalNoteList,
 } from './Journal/journal';
 
-export function book(info: BookInterface, pages: string[]): {bibliography: string, notes: string[]} {
+export function book(info: BookInterface, pages: Array<{page: string}>): {bibliography: string, notes: string[]} {
   return { bibliography: bookBibliography(info), notes: bookNoteList(info, pages) };
 }
 
-export function eBook(info: BookInterface, pages: string[]): {bibliography: string, notes: string[]} {
+export function eBook(info: BookInterface, pages: Array<{page: string}>): {bibliography: string, notes: string[]} {
   if (!info.url) throw new Error('URL not specified');
   return { bibliography: eBookBibliography(info, info.url), notes: eBookNoteList(info, pages, info.url) };
 }
 
-export function journal(info: JournalInterface, pages: string[]): {bibliography: string, notes: string[]} {
+export function journal(info: JournalInterface, pages: Array<{page: string}>): {bibliography: string, notes: string[]} {
   return { bibliography: journalBibliography(info), notes: journalNoteList(info, pages) };
 }
 
-export function eJournal(info: JournalInterface, pages: string[]): {bibliography: string, notes: string[]} {
+export function eJournal(info: JournalInterface, pages: Array<{page: string}>): {bibliography: string, notes: string[]} {
   if (!info.url) throw new Error('URL not specified');
   return { bibliography: eJournalBibliography(info, info.url), notes: eJournalNoteList(info, pages, info.url) };
 }