+++ test/cmosCitation.test.ts
@@ -15,20 +15,20 @@ describe("createBookCitation() testing", () => {
             publisher: "Charles Scribner's Sons",
             placeOfPublication: "New York",
             yearOfPublication: "2000"
-        }, ["1", "2", "3"]);
+        }, [{page: "1"}, {page: "2"}, {page: "3"}]);
 
         expect(citation).toStrictEqual({
             "bibliography": "Fitzgerald, F. Scott. The Great Gatsby. New York: Charles Scribner's Sons, 2000.",
             "notes": ["F. Scott Fitzgerald, The Great Gatsby (New York: Charles Scribner's Sons, 2000), 1.",
                 "Fitzgerald, 2.",
                 "Fitzgerald, 3."]
-        })
+        });
     });
 
     test("should create book citation with only title", () => {
         let citation = book({
             title: "The Great Gatsby"
-        }, ["1", "2", "3"]);
+        }, [{page: "1"}, {page: "2"}, {page: "3"}]);
 
         expect(citation).toStrictEqual({
             "bibliography": "The Great Gatsby.",
@@ -36,6 +36,22 @@ describe("createBookCitation() testing", () => {
                 "The Great Gatsby, 1.",
                 "The Great Gatsby, 2.",
                 "The Great Gatsby, 3."]
+        });
+    });
+
+    test("should create note citations with invalid pages", () => {
+        let citation = book({
+            title: "The Great Gatsby",
+            authorList: [{first: "F. Scott", last: "Fitzgerald"}],
+            publisher: "Charles Scribner's Sons",
+            placeOfPublication: "New York",
+            yearOfPublication: "2000"
+        }, [{page: "1"}, {page: ""}, {page: "2"}]);
+
+        expect(citation).toStrictEqual({
+            "bibliography": "Fitzgerald, F. Scott. The Great Gatsby. New York: Charles Scribner's Sons, 2000.",
+            "notes": ["F. Scott Fitzgerald, The Great Gatsby (New York: Charles Scribner's Sons, 2000), 1.",
+                "Fitzgerald, 2."]
         });
     });
 });
@@ -49,7 +65,7 @@ describe("createEBookCitation() testing", () => {
             placeOfPublication: "New York",
             yearOfPublication: "2000",
             url: "www.thegreatgatsby.com"
-        }, ["1"]);
+        }, [{page: "1"}]);
 
         expect(citation).toStrictEqual({
             "bibliography": "Fitzgerald, F. Scott. The Great Gatsby. New York: Charles Scribner's Sons, 2000. www.thegreatgatsby.com.",
@@ -61,7 +77,7 @@ describe("createEBookCitation() testing", () => {
         let citation = eBook({
             title: "The Great Gatsby",
             url: "www.thegreatgatsby.com"
-        }, ["1"]);
+        }, [{page: "1"}]);
 
         expect(citation).toStrictEqual({
             "bibliography": "The Great Gatsby. www.thegreatgatsby.com.",
@@ -71,7 +87,7 @@ describe("createEBookCitation() testing", () => {
 
     test("should throw error if missing url field", () => {
         try {
-            eBook({title: "The Great Gatsby"}, ["1"]);
+            eBook({title: "The Great Gatsby"}, [{page: "1"}]);
             fail("No error thrown with missing URL field");
         } catch (e) {
             expect(e.message).toBe("URL not specified");
@@ -84,7 +100,7 @@ describe("createJournalCitation() testing", () => {
         let citation = journal({
             title: "Title", authorList: [{first: "First", last: "Last"}], publisher: "Publisher",
             yearOfPublication: "Year", volume: "vol", issue: "iss", startRange: "1", endRange: "10"
-        }, ["1", "2", "3"]);
+        }, [{page: "1"}, {page: "2"}, {page: "3"}]);
         expect(citation).toStrictEqual({
             "bibliography": "Last, First. \"Title.\" Publisher vol, no. iss (Year): 1-10.",
             "notes": [
@@ -98,7 +114,7 @@ describe("createJournalCitation() testing", () => {
     test("should create journal citation with only title", () => {
         let citation = journal({
             title: "Title"
-        }, ["1", "2", "3"]);
+        }, [{page: "1"}, {page: "2"}, {page: "3"}]);
 
         expect(citation).toStrictEqual({
             "bibliography": "\"Title.\"",
@@ -117,7 +133,7 @@ describe("createEJournalCitation() testing", () => {
             title: "Title", authorList: [{first: "First", last: "Last"}], publisher: "Publisher",
             yearOfPublication: "Year", volume: "vol", issue: "iss", startRange: "1", endRange: "10",
             url: "www.test.com"
-        }, ["1"]);
+        }, [{page: "1"}]);
         expect(citation).toStrictEqual({
             "bibliography": "Last, First. \"Title.\" Publisher vol, no. iss (Year): 1-10. www.test.com.",
             "notes": ["First Last, \"Title,\" Publisher vol, no. iss (Year): 1, www.test.com."]
@@ -128,7 +144,7 @@ describe("createEJournalCitation() testing", () => {
         let citation = eJournal({
             title: "Title",
             url: "www.test.com"
-        }, ["1"]);
+        }, [{page: "1"}]);
 
         expect(citation).toStrictEqual({
             "bibliography": "\"Title.\" www.test.com.",
@@ -138,7 +154,7 @@ describe("createEJournalCitation() testing", () => {
 
     test("should throw error if missing url field", () => {
         try {
-            eJournal({title: "The Great Gatsby"}, ["1"]);
+            eJournal({title: "The Great Gatsby"}, [{page: "1"}]);
             expect(true).toBe(false);   // Will trigger if no error thrown
         } catch (e) {
             expect(e.message).toBe("URL not specified");